Refactored submodules to be private
Fixes #174.
