+++ cbor2/__init__.py
@@ -1,27 +1,36 @@
-from .decoder import CBORDecoder as CBORDecoder
-from .decoder import load as load
-from .decoder import loads as loads
-from .encoder import CBOREncoder as CBOREncoder
-from .encoder import dump as dump
-from .encoder import dumps as dumps
-from .encoder import shareable_encoder as shareable_encoder
-from .types import CBORDecodeEOF as CBORDecodeEOF
-from .types import CBORDecodeError as CBORDecodeError
-from .types import CBORDecodeValueError as CBORDecodeValueError
-from .types import CBOREncodeError as CBOREncodeError
-from .types import CBOREncodeTypeError as CBOREncodeTypeError
-from .types import CBOREncodeValueError as CBOREncodeValueError
-from .types import CBORError as CBORError
-from .types import CBORSimpleValue as CBORSimpleValue
-from .types import CBORTag as CBORTag
-from .types import undefined as undefined
+from typing import Any
+
+from ._decoder import CBORDecoder as CBORDecoder
+from ._decoder import load as load
+from ._decoder import loads as loads
+from ._encoder import CBOREncoder as CBOREncoder
+from ._encoder import dump as dump
+from ._encoder import dumps as dumps
+from ._encoder import shareable_encoder as shareable_encoder
+from ._types import CBORDecodeEOF as CBORDecodeEOF
+from ._types import CBORDecodeError as CBORDecodeError
+from ._types import CBORDecodeValueError as CBORDecodeValueError
+from ._types import CBOREncodeError as CBOREncodeError
+from ._types import CBOREncodeTypeError as CBOREncodeTypeError
+from ._types import CBOREncodeValueError as CBOREncodeValueError
+from ._types import CBORError as CBORError
+from ._types import CBORSimpleValue as CBORSimpleValue
+from ._types import CBORTag as CBORTag
+from ._types import FrozenDict as FrozenDict
+from ._types import undefined as undefined
 
 try:
     from _cbor2 import *  # noqa: F403
 except ImportError:
     # Couldn't import the optimized C version; ignore the failure and leave the
     # pure Python implementations in place.
-    pass
+
+    # Re-export imports so they look like they live directly in this package
+    key: str
+    value: Any
+    for key, value in list(locals().items()):
+        if callable(value) and getattr(value, "__module__", "").startswith("cbor2."):
+            value.__module__ = __name__
 else:
     # The pure Python implementations are replaced with the optimized C
     # variants, but we still need to create the encoder dictionaries for the C
@@ -32,8 +41,8 @@ else:
 
         import _cbor2
 
-        from .encoder import canonical_encoders, default_encoders
-        from .types import CBORSimpleValue, CBORTag, undefined  # noqa: F811
+        from ._encoder import canonical_encoders, default_encoders
+        from ._types import CBORSimpleValue, CBORTag, undefined  # noqa: F811
 
         _cbor2.default_encoders = OrderedDict(
             [
