--- docs/modules/decoder.rst
@@ -1,5 +1,0 @@
-:mod:`cbor2.decoder`
-====================
-
-.. automodule:: cbor2.decoder
-    :members:
