--- docs/modules/encoder.rst
@@ -1,5 +1,0 @@
-:mod:`cbor2.encoder`
-====================
-
-.. automodule:: cbor2.encoder
-    :members:
