--- docs/modules/types.rst
@@ -1,5 +1,0 @@
-:mod:`cbor2.types`
-==================
-
-.. automodule:: cbor2.types
-    :members:
