+++ docs/usage.rst
@@ -1,6 +1,8 @@
 Basic usage
 ===========
 
+.. py:currentmodule:: cbor2
+
 Serializing and deserializing with cbor2 is pretty straightforward::
 
     from cbor2 import dumps, loads, load
@@ -27,12 +29,12 @@ Date/time handling
 The CBOR specification does not support naïve datetimes (that is, datetimes where ``tzinfo`` is
 missing). When the encoder encounters such a datetime, it needs to know which timezone it belongs
 to. To this end, you can specify a default timezone by passing a :class:`~datetime.tzinfo` instance
-to :func:`~cbor2.encoder.dump`/:func:`~cbor2.encoder.dumps` call as the ``timezone`` argument.
+to :func:`dump`/:func:`dumps` call as the ``timezone`` argument.
 Decoded datetimes are always timezone aware.
 
 By default, datetimes are serialized in a manner that retains their timezone offsets. You can
 optimize the data stream size by passing ``datetime_as_timestamp=False`` to
-:func:`~cbor2.encoder.dump`/:func:`~cbor2.encoder.dumps`, but this causes the timezone offset
+:func:`dump`/:func:`dumps`, but this causes the timezone offset
 information to be lost.
 
 In versions prior to 4.2 the encoder would convert a ``datetime.date`` object into a
@@ -51,11 +53,11 @@ Cyclic (recursive) data structures
 ----------------------------------
 
 If the encoder encounters a shareable object (ie. list or dict) that it has seen before, it will
-by default raise :exc:`~cbor2.types.CBOREncodeError` indicating that a cyclic reference has been
+by default raise :exc:`CBOREncodeError` indicating that a cyclic reference has been
 detected and value sharing was not enabled. CBOR has, however, an extension specification that
 allows the encoder to reference a previously encoded value without processing it again. This makes
 it possible to serialize such cyclic references, but value sharing has to be enabled by passing
-``value_sharing=True`` to :func:`~cbor2.encoder.dump`/:func:`~cbor2.encoder.dumps`.
+``value_sharing=True`` to :func:`dump`/:func:`dumps`.
 
 .. warning:: Support for value sharing is rare in other CBOR implementations, so think carefully
     whether you want to enable it. It also causes some line overhead, as all potentially shareable
@@ -65,7 +67,7 @@ String references
 -----------------
 
 When ``string_referencing=True`` is passed to
-:func:`~cbor2.encoder.dump`/:func:`~cbor2.encoder.dumps`, if the encoder would encode a string that
+:func:`dump`/:func:`dumps`, if the encoder would encode a string that
 it has previously encoded and where a reference would be shorter than the encoded string, it
 instead encodes a reference to the nth sufficiently long string already encoded.
 
@@ -90,7 +92,7 @@ Tag   Semantics                                Python type(s)
 28    Mark shared value                        N/A
 29    Reference shared value                   N/A
 30    Rational number                          fractions.Fraction
-35    Regular expression                       ``_sre.SRE_Pattern`` (result of ``re.compile(...)``)
+35    Regular expression                       re.Pattern (result of ``re.compile(...)``)
 36    MIME message                             email.message.Message
 37    Binary UUID                              uuid.UUID
 256   String reference namespace               N/A
@@ -100,10 +102,10 @@ Tag   Semantics                                Python type(s)
 55799 Self-Described CBOR                      object
 ===== ======================================== ====================================================
 
-Arbitary tags can be represented with the :class:`~cbor2.types.CBORTag` class.
+Arbitary tags can be represented with the :class:`CBORTag` class.
 
 If you want to write a file that is detected as CBOR by the Unix ``file`` utility, wrap your data in
-a ``~cbor2.types.CBORTag`` object like so::
+a :class:`CBORTag` object like so::
 
     from cbor2 import dump, CBORTag
